fix windows build/convolution, see new comment (main.cpp) for build instructions

+++ src/Image.cpp
@@ -92,36 +92,28 @@ Image &Image::convolve_clamp_to_0(uint8_t channel, uint32_t ker_w, uint32_t ker_
 	return *this;
 }
 
-Image &Image::convolve_clamp_to_border(uint8_t channel, uint32_t ker_w, uint32_t ker_h, double ker[], uint32_t cr, uint32_t cc)
-{
-	uint8_t *new_data = new uint8_t[w * h];
-	uint64_t center = cr * ker_w + cc;
-	for (uint64_t k = channel; k < size; k += channels)
-	{
+Image& Image::convolve_clamp_to_border(uint8_t channel, uint32_t ker_w, uint32_t ker_h, double ker[], uint32_t cr, uint32_t cc) {
+	uint8_t* new_data = new uint8_t[w * h];
+	uint64_t center = (size_t)cr * ker_w + cc;
+	for (uint64_t k = channel; k < size; k += channels) {
 		double c = 0;
-		for (long i = -((long)cr); i < (long)ker_h - cr; ++i)
-		{
+		for (long i = -((long)cr); i < (long)ker_h - (long)cr; ++i) {
 			long row = ((long)k / channels) / w - i;
-			if (row < 0)
-			{
+			if (row < 0) {
 				row = 0;
 			}
-			else if (row > h - 1)
-			{
+			else if (row > h - 1) {
 				row = h - 1;
 			}
-			for (long j = -((long)cc); j < (long)ker_w - cc; ++j)
-			{
+			for (long j = -((long)cc); j < (long)ker_w - (long)cc; ++j) {
 				long col = ((long)k / channels) % w - j;
-				if (col < 0)
-				{
+				if (col < 0) {
 					col = 0;
 				}
-				else if (col > w - 1)
-				{
+				else if (col > w - 1) {
 					col = w - 1;
 				}
-				c += ker[center + i * (long)ker_w + j] * data[(row * w + col) * channels + channel];
+				c += ker[center + (long long)i * (long)ker_w + j] * data[(row * w + col) * channels + channel];
 			}
 		}
 		new_data[k / channels] = BYTE_BOUND(c);
